Add autosync functionality
org-caldav-autosync-mode: Minor mode that syncs with CalDav when
idling, after saving changes to an org-caldav file.  Purpose is to
auto propagates changes from Org to CalDav.

org-caldav-agenda-maybe-sync: Drop-in replacement for org-agenda, that
occasionally asks to sync with CalDav before building the agenda.
Purpose is to auto propagate changes from CalDav into Org.